backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]>=0.34.0
google-genai>=1.0.0
google-adk>=1.0.0
pydantic>=2.12.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
sqlmodel>=0.0.21
httpx>=0.28.0
numpy>=2.0.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
